refactor: move visual comparison tool from test to example
Move `generate_comparison_images` from tests/compare_backends.rs to
examples/compare_backends.rs. This was a slow, manual-inspection-only
tool that shouldn't run in CI. Now available via:

  cargo run --example compare_backends

The automated `cross_backend_dimension_parity` test stays in the test
file and runs normally with `cargo test`.
