+++ .github/workflows/test.yml
@@ -25,7 +25,7 @@ jobs:
         run: cargo test --all-features
 
       - name: Run ignored tests (require ImageMagick)
-        run: cargo test --all-features --bin simple-gal --test compare_backends -- --ignored
+        run: cargo test --all-features --bin simple-gal -- --ignored
 
       - name: Build release binary
         run: cargo build --release
